% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_v_exp_Q_t}
\alias{cpp_v_exp_Q_t}
\title{Evaluate matrix exponential action using uniformization}
\usage{
cpp_v_exp_Q_t(v, Q, t = 1, tolerance = 1e-12)
}
\arguments{
\item{v}{A row vector}

\item{Q}{a sparse rate matrix}

\item{t}{Numeric time value, action evaluated is v'exp(Q*t)}

\item{tolerance}{Maximum error of the numeric calculation}
}
\description{
Evaluate matrix exponential action using uniformization
}
