% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{expmAction-package}
\alias{expmAction}
\alias{expmAction-package}
\title{expmAction: Evaluates v'exp(Q) for sparse matrix Q}
\description{
An efficient implementation of the uniformization algorithm is used to evaluate the matrix exponential action v'exp(Q) for a positive vector v.
}
\keyword{internal}
