% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{expmAction-package}
\alias{expmAction}
\alias{expmAction-package}
\title{expmAction: What the package does (short line)}
\description{
More about what it does (maybe more than one line)
}
\keyword{internal}
